+++ modules/yup_dsp/yup_dsp.cpp
@@ -83,7 +83,8 @@
 #include "frequency/yup_SpectrumAnalyzerState.cpp"
 #include "onsets/yup_Spectrogram.cpp"
 #include "onsets/yup_FilterBank.cpp"
-#include "onsets/yup_OnsetDetectionFunction.cpp"
+#include "onsets/yup_SuperFluxODF.cpp"
+#include "onsets/yup_ComplexFluxODF.cpp"
 #include "onsets/yup_OnsetPeakPicker.cpp"
 #include "onsets/yup_OnsetDetector.cpp"
 #include "metering/yup_LevelProcessor.cpp"
